Making comparison networks @jameals
Made three new plots, first is coastwide with edge width proportional to
correlation (grey is negative, blue is positive). Second is the jaquard
version of the same network and nodes are in same place to facilitate
comparision. Third is the full coastwide jacquard network (without
masking to drop pairwise interactions among fisheries with < 10 vessels
doing both).